+++ src/js/add-to-favorites.js
@@ -1,0 +1,21 @@
+import favoritesArray from "./favorites-array";
+import { dataRecipes } from "./all_recipes";
+import { save } from "./api";
+
+function addToFavorites(event){
+    const recipeId = event.target.closest('.card-template').dataset.id;
+    favoriteRecipe = findRecipe(recipeId);
+  }
+
+function findRecipe(recipeId){
+   const item = dataRecipes.find(({ _id }) => _id === recipeId)
+   if (favoritesArray.find(({_id}) => _id === recipeId) === undefined){
+    favoritesArray.push(item)
+    save("KEY_FAVORITE", favoritesArray)
+   }
+   else{
+    removeFromFavorites()
+   }
+}
+
+export {addToFavorites}
