+++ src/js/favorites-array.js
@@ -1,0 +1,2 @@
+const favoritesArray = [];
+export default favoritesArray;
